添加注释;将梯形积分部分用类重新实现了一遍

+++ lib/radians.m
@@ -1,6 +1,6 @@
 function rad = radians(deg)
 
-    rad = deg/57.3;
+    rad = double(deg)/57.3;
 
 end
